requirements_test.txt
-r requirements.txt

# Development Testing
pytest>=8.0.0
pytest-homeassistant-custom-component>=0.13.236
pytest-cov>=4.0.0
